updated upload dialog to include single occupancy

+++ app/src/main/java/com/hmmelton/releave/data/models/RestroomRequestBody.kt
@@ -11,5 +11,5 @@ data class RestroomRequestBody(
     val location: String,
     @Json(name = "is_locked") val isLocked: Boolean = true,
     @Json(name = "is_single_occupancy") val isSingleOccupancy: Boolean = false,
-    val rating: Double
+    val rating: Int
 )
